pyproject.toml
[project]
name = "mytoolbox"
version = "0.1.0"
description = "基于OpenAI Whisper的语音转文字工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
    "faster-whisper>=1.0.0",
    "psutil",
    "torch",
    "torchaudio",
    "PyYAML",
]

[project.optional-dependencies]
gui = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
tts = [
    "mycroft-mimic3-tts>=0.2.4",
]
onnx = [
    "onnxruntime>=1.16.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
]

[tool.setuptools.packages.find]
include = ["core*", "config*", "modules*", "web*", "whisper_engines*"]
